completed all 15 tasks given by Jayesh on 29/8/19
