+++ raft-store/snapshots/2-3-1567095041952/meta.json
@@ -1,0 +1,1 @@
+{"Version":1,"ID":"2-3-1567095041952","Index":3,"Term":2,"Peers":"kbIxOTIuMTY4LjAuMTAwOjQyMzQ=","Configuration":{"Servers":[{"Suffrage":0,"ID":"192.168.0.100","Address":"192.168.0.100:4234"}]},"ConfigurationIndex":1,"Size":1017,"CRC":"E6lIB1AWxyw="}
