+++ raft-store/snapshots/2-6-1567097322971/meta.json
@@ -1,0 +1,1 @@
+{"Version":1,"ID":"2-6-1567097322971","Index":6,"Term":2,"Peers":"kbIxOTIuMTY4LjAuMTAwOjQyMzQ=","Configuration":{"Servers":[{"Suffrage":0,"ID":"192.168.0.100","Address":"192.168.0.100:4234"}]},"ConfigurationIndex":1,"Size":1017,"CRC":"E6lIB1AWxyw="}
