--- src/pages/database/style.css
@@ -1,155 +1,0 @@
-.rules-main-wrapper {
-  box-shadow: 0px 2px 12px rgba(0, 0, 0, 0.04), 0px 6px 12px rgba(0, 0, 0, 0.0125), 0px 12px 24px rgba(0, 0, 0, 0.00625), 0px 24px 48px rgba(0, 0, 0, 0.003125), 0px 48px 96px rgba(0, 0, 0, 0.003125);
-  margin-top: 40px;
-}
-
-.rules-main-wrapper .rulesTable {
-  border: 1px solid #E4E4E4;
-  height: 300px;
-  overflow-y: scroll;
-
-}
-
-::-webkit-scrollbar {
-  width: 8px;
-  border-radius: 2px;
-}
-
-::-webkit-scrollbar-track {
-  background: transparent;
-}
-
-::-webkit-scrollbar-thumb {
-  background: #E0E0E0;
-}
-
-.rules-main-wrapper .rulesTable .rule {
-  padding: 20px;
-  border: 1px solid #E0E0E0;
-  color: #000;
-  font-weight: 600;
-}
-
-.rules-main-wrapper .rulesTable .selected {
-  border-left: 5px solid #FF6700;
-
-}
-
-.rules-main-wrapper {
-    color: black;
-  padding: 20px;
-}
-
-.addaRule {
-    padding: 20px;
-    color: black;
-    border: 1px solid #E4E4E4;
-}
-
-.rules-main-wrapper .rulesTable .addaRule .addIcon {
-  margin-right: 20px
-}
-
-.rules-main-wrapper .code .code-hint {
-  padding: 20px;
-  border: 1px solid #E4E4E4;
-}
-
-.rules-main-wrapper .code .code-mirror {
-  height: 300px;
-}
-
-.rules-main-wrapper .add-a-rule {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  cursor: pointer;
-}
- .add-a-rule   {
-  cursor: pointer;
-}
-
-.collections {
-  font-style: normal;
-  font-weight: bold;
-  font-size: 20px;
-  line-height: 27px;
-
-  color: #000000;
-}
-
-.tablehead {
-  color: black;
-  text-align: center;
-  font-style: normal;
-  font-weight: 600;
-  font-size: 20px;
-  line-height: 27px;
-  padding: 12px 0;
-  border: 1px solid #e4e4e4;
-}
-
-.tabledata {
-  font-style: normal;
-  font-weight: 600;
-  font-size: 16px;
-  line-height: 22px;
-  text-align: center;
-  color: black;
-  padding: 12px 20px;
-  border-bottom: 1px solid #e4e4e4;
-  border-left: 1px solid #e4e4e4;
-  border-right: 1px solid #e4e4e4;
-  list-style: none;
-  cursor: pointer;
-}
-.activedata {
-  background: #ecf6fd;
-  border-left: 5px solid #ff6700;
-  padding: 12px 20px;
-}
-
-.activedatabackground {
-  background: #ecf6fd;
-}
-
-.codebox {
-  font-style: normal;
-  font-weight: normal;
-  font-size: 14px;
-  line-height: 19px;
-  padding: 16px 0;
-  border-right: 1px solid #e4e4e4;
-  border-top: 1px solid #e4e4e4;
-  border-bottom: 1px solid #e4e4e4;
-  color: rgba(3, 27, 78, 0.7);
-}
-
-.ant-tabs-tab {
-  font-style: normal;
-  font-weight: 600;
-  font-size: 18px;
-  line-height: 25px;
-  color: black;
-}
-
-.ant-tabs-tab-active {
-  font-style: normal;
-  font-weight: 600;
-  font-size: 18px;
-  line-height: 25px;
-  color: black;
-}
-
-.db-page-content {
-
-  position: absolute;
-  top: 51px;
-  left: 250px;
-  right: 0;
-}
-
-.rules-schema-table {
-
-  padding: 0px 120px 0 56px;
-}
